Create module for Cvx Mint Amount in utils
